+++ src/lte/model/lte-phy.h
@@ -141,7 +141,7 @@ public:
 
 
   /**
-  * \returns the RB gruop size according to the bandwidth
+  * \returns the RB group size according to the bandwidth
   */
   uint8_t GetRbgSize (void) const;
   
@@ -266,13 +266,13 @@ protected:
    * The UL bandwidth in number of PRBs.
    * Specified by the upper layer through CPHY SAP.
    */
-  uint8_t m_ulBandwidth;
+  uint16_t m_ulBandwidth;
   /**
    * The DL bandwidth in number of PRBs.
    * Specified by the upper layer through CPHY SAP.
    */
-  uint8_t m_dlBandwidth;
-  /// The RB gruop size according to the bandwidth.
+  uint16_t m_dlBandwidth;
+  /// The RB group size according to the bandwidth.
   uint8_t m_rbgSize;
   /**
    * The downlink carrier frequency.
